# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Water-id Translator Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@water-id.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://translator.local/api/v1",
            "description": "Water-id Translator API Server"
        }
    ],
    "paths": {
...
        "/phraseCategories": {
            "get": {
                "tags": [
                    "PhraseCategories"
                ],
                "summary": "Get list of phraseCategories",
                "description": "Returns list of phraseCategories",
                "operationId": "getPhraseCategoriesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of item per page",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Column name to sort by",
                        "schema": {
                            "type": "string"
                        },
                        "example": "name"
                    },
                    {
                        "name": "sortDesc",
                        "in": "query",
                        "description": "Column name to sort descending",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search Query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phraseCategory",
                        "in": "query",
                        "description": "Phrase Category Name",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseCategoryResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PhraseCategories"
                ],
                "summary": "Store new phraseCategory",
                "description": "Returns phraseCategory data",
                "operationId": "storePhraseCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePhraseCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseCategoryOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnprocessableEntityException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/phraseCategories/{id}": {
            "get": {
                "tags": [
                    "PhraseCategories"
                ],
                "summary": "Get phraseCategory information",
                "description": "Returns phraseCategory data",
                "operationId": "getPhraseCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PhraseCategory id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseCategoryOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "PhraseCategories"
                ],
                "summary": "Update existing phraseCategory",
                "description": "Returns updated phraseCategory data",
                "operationId": "updatePhraseCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PhraseCategory id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePhraseCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhraseCategoryOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRequestException"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "PhraseCategories"
                ],
                "summary": "Delete existing phraseCategory",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePhraseCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "PhraseCategory id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiUnAuthException"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid scope(s) provided.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiAccessDeniedException"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiNotFoundException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiAccessDeniedException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "AccessDenied"
                    }
                },
                "type": "object"
            },
            "ApiNotFoundException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "ApiRequestException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Bad Request"
                    }
                },
                "type": "object"
            },
...
            "ApiUnAuthException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
            "ApiUnprocessableEntityException": {
                "properties": {
                    "message": {
                        "description": "The err message",
                        "type": "string",
                        "example": "Unprocessable Entity"
                    }
                },
                "type": "object"
            },
...
            "StorePhraseCategoryRequest": {
                "title": "Store Phrase Category request",
                "description": "Store Phrase Category request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "example": "A new phrase category name"
                    },
                    "filename": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string",
                        "example": "A phrase category description"
                    }
                },
                "type": "object"
            },
...
            "UpdatePhraseCategoryRequest": {
                "title": "Update phrase category request",
                "description": "Update phrase category request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "example": "A new phrase phrase category name"
                    },
                    "filename": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string",
                        "example": "A phrase category description"
                    }
                },
                "type": "object"
            },
...
            "PhraseCategoryResource": {
                "title": "PhraseCategoryResource",
                "description": "Phrase Category resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PhraseCategoryResource"
                }
            },
...
            "PhraseCategoryOut": {
                "title": "PhraseCategoryOut",
                "description": "PhraseCategoryOut model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "format": "string",
                        "example": "Notes"
                    },
                    "filename": {
                        "title": "File Name",
                        "description": "File Name",
                        "type": "integer",
                        "format": "string",
                        "example": "[{'@attributes':{'filename':'..\\/flux\\/views\\/settings\\/dataschemes\\/AccountScheme.qml','line':'26'}}]"
                    },
                    "phrases_count": {
                        "title": "Phrase Count",
                        "description": "Phrases Count",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PhraseCategoryOut"
                }
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Enter \"Bearer + [space] + [token]\" in the text input below(Value).",
                "example": "Bearer 1qaz2wsx3edc",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
...
}